폴더 구조 리팩토링 -types 폴더 분리, 파일 통합 -useClient 파일 이동(lib -> constants) -불필요한 css 파일 삭제
