+++ src/components/layout/SideNavigation.tsx
@@ -4,7 +4,7 @@ import React, { useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import { env } from 'process'
 import LocalLibraryRoundedIcon from '@mui/icons-material/LocalLibraryRounded'
-import { Box, Button, Divider, ListItemButton, ListItemText, MenuItem, MenuList, ThemeProvider } from '@/lib/useClient/mui'
+import { Box, Button, Divider, ListItemButton, ListItemText, MenuItem, MenuList, ThemeProvider } from '@/constants/useClient/mui'
 import { theme } from '@/assets/styles/theme'
 import GoogleIcon from '@/assets/image/button/btn_google_light_normal_ios.svg'
 import { useAuth } from '@/hooks/useAuth'
