--- src/css/layout.css
@@ -1,1 +1,0 @@
-html,body{display:flex;flex-direction:row;align-items:center;max-width:100vw;overflow-x:hidden;height:100%;min-height:100%;margin:0px}
