--- src/css/layout.module.css
@@ -1,1 +1,0 @@
-.body{max-width:100vw;overflow-x:hidden;display:flex;flex-direction:row;align-items:center;height:100%;margin:0px}
