--- src/lib/types/type.ts
@@ -1,7 +1,0 @@
-type Book = {
-  id: string
-  isbn: string
-  imgUrl: string
-  title: string
-  author: string
-}
