+++ src/types/type.ts
@@ -5,3 +5,11 @@ export interface Member {
   profileTxt?: string
   createDate?: Date
 }
+
+type Book = {
+  id: string
+  isbn: string
+  imgUrl: string
+  title: string
+  author: string
+}
